criando classe para erros personalizados

+++ src/storage/group/groupCreate.ts
@@ -1,10 +1,17 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "../storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
+import { AppError } from "@utils/appError";
 
 export const groupCreate = async (newGroup: string) => {
     try {
         const storedGroups = await groupsGetAll();
+
+        const groupAlreadyExists = storedGroups.includes(newGroup);
+
+        if(groupAlreadyExists) {
+            throw new AppError('Group already exists.')
+        }
 
         const storage = JSON.stringify([...storedGroups, newGroup]);
         await AsyncStorage.setItem(GROUP_COLLECTION, storage);
